Move BinaryWriterWrapper to the correct namespace
