+++ SAGESharp.Tests/IO/Binary/BinaryWriterWrapperTests.cs
@@ -7,10 +7,11 @@ using FluentAssertions;
 using NUnit.Framework;
 using NUtils.Extensions;
 using SAGESharp.IO;
+using SAGESharp.IO.Binary;
 using System;
 using System.IO;
 
-namespace SAGESharp.Tests.IO
+namespace SAGESharp.Tests.IO.Binary
 {
     class BinaryWriterWrapperTests
     {
