+++ SAGESharp.Tests/SAGESharp.Tests.csproj
@@ -83,8 +83,8 @@
     <Compile Include="IO\BinarySerializableSerializerTests.cs" />
     <Compile Include="IO\BinarySerializerTests.cs" />
     <Compile Include="IO\BinaryWriterTests.cs" />
-    <Compile Include="IO\BinaryWriterWrapperTests.cs" />
     <Compile Include="IO\Binary\BinaryReaderWrapperTests.cs" />
+    <Compile Include="IO\Binary\BinaryWriterWrapperTests.cs" />
     <Compile Include="IO\Binary\ReaderTests.cs" />
     <Compile Include="IO\EdgeTests.cs" />
     <Compile Include="IO\ListNodeTests.cs" />
